feat: confirm before deleting

+++ src/views/JobView.vue
@@ -20,9 +20,14 @@ const state = reactive({
 
 const deleteJob = async () => {
     try {
-        await axios.delete(`/api/jobs/${jobId}`);
-        toast.success('Job deleted successfully');
-        router.push('/jobs');
+        const confirm = window.confirm(
+            'Are you sure you want to delete this job?'
+        );
+        if(confirm){
+            await axios.delete(`/api/jobs/${jobId}`);
+            toast.success('Job deleted successfully');
+            router.push('/jobs');
+        }
     } catch (error) {
         toast.error('Error deleting job');
     }
